+++ lib/jsdom/living/helpers/selectors.js
@@ -3,7 +3,7 @@
 const domSelector = require("@asamuzakjp/dom-selector");
 const DOMException = require("domexception/webidl2js-wrapper");
 
-exports.matchesWithOutThrow = (selectors, element) => {
+exports.matchesDontThrow = (selectors, element) => {
   let matched;
   try {
     matched = domSelector.matches(selectors, element);
