+++ lib/jsdom/living/helpers/style-rules.js
@@ -2,7 +2,8 @@
 const cssom = require("rrweb-cssom");
 const { CSSStyleDeclaration } = require("cssstyle");
 const defaultStyleSheet = require("../../browser/default-stylesheet");
-const { matchesWithOutThrow } = require("./selector.js");
+const idlUtils = require("../generated/utils");
+const { matchesDontThrow } = require("./selectors.js");
 
 const { forEach, indexOf } = Array.prototype;
 
@@ -101,14 +102,8 @@ exports.getDeclarationForElement = elementImpl => {
   return declaration;
 };
 
-function matches(rule, element) {
-  let matched;
-  try {
-    matched = matchesWithOutThrow(rule.selectorText, element);
-  } catch (e) {
-    matched = false;
-  }
-  return matched;
+function matches(rule, elementImpl) {
+  return matchesDontThrow(rule.selectorText, idlUtils.wrapperForImpl(elementImpl));
 }
 
 // Naive implementation of https://drafts.csswg.org/css-cascade-4/#cascading
