+++ lib/jsdom/living/nodes/Document-impl.js
@@ -12,7 +12,6 @@ const whatwgURL = require("whatwg-url");
 const StyleSheetList = require("../generated/StyleSheetList.js");
 const { domSymbolTree } = require("../helpers/internal-constants");
 const eventAccessors = require("../helpers/create-event-accessor");
-const { querySelector, querySelectorAll } = require("../helpers/selector");
 const { asciiLowercase, stripAndCollapseASCIIWhitespace } = require("../helpers/strings");
 const { childTextContent } = require("../helpers/text");
 const { HTML_NS, SVG_NS } = require("../helpers/namespaces");
@@ -88,11 +87,6 @@ function toLastModifiedString(date) {
     " " + pad(date.getHours()) +
     ":" + pad(date.getMinutes()) +
     ":" + pad(date.getSeconds());
-}
-
-function shouldAlwaysSelectNothing(elImpl) {
-  // This is true during initialization.
-  return elImpl === elImpl._ownerDocument && !elImpl.documentElement;
 }
 
 const eventInterfaceTable = {
@@ -915,16 +909,6 @@ class DocumentImpl extends NodeImpl {
     copy._URL = this._URL;
     copy._origin = this._origin;
     return copy;
-  }
-
-  querySelector(selectors) {
-    return querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-  }
-
-  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
-  querySelectorAll(selectors) {
-    const nodes = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-    return NodeList.create(this._globalObject, [], { nodes });
   }
 }
