+++ lib/jsdom/living/nodes/DocumentFragment-impl.js
@@ -1,12 +1,10 @@
 "use strict";
 const { mixin } = require("../../utils");
 const { domSymbolTree } = require("../helpers/internal-constants");
-const { querySelector, querySelectorAll } = require("../helpers/selector");
 const NODE_TYPE = require("../node-type");
 const NodeImpl = require("./Node-impl").implementation;
 const NonElementParentNodeImpl = require("./NonElementParentNode-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
-const NodeList = require("../generated/NodeList");
 const idlUtils = require("../generated/utils");
 
 class DocumentFragmentImpl extends NodeImpl {
@@ -35,15 +33,6 @@ class DocumentFragmentImpl extends NodeImpl {
     }
 
     return null;
-  }
-
-  querySelector(selectors) {
-    return querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-  }
-
-  querySelectorAll(selectors) {
-    const nodes = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-    return NodeList.create(this._globalObject, [], { nodes });
   }
 }
