+++ lib/jsdom/living/nodes/Element-impl.js
@@ -1,6 +1,7 @@
 "use strict";
 const { HTML_NS } = require("../helpers/namespaces");
 const { mixin, memoizeQuery } = require("../../utils");
+const idlUtils = require("../generated/utils");
 const NodeImpl = require("./Node-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
 const ChildNodeImpl = require("./ChildNode-impl").implementation;
@@ -11,12 +12,10 @@ const { parseFragment } = require("../../browser/parser");
 const InnerHTMLImpl = require("../domparsing/InnerHTML-impl").implementation;
 const { fragmentSerialization } = require("../domparsing/serialization");
 const { domSymbolTree } = require("../helpers/internal-constants");
-const { closest, matches, querySelector, querySelectorAll } = require("../helpers/selector");
+const { closest, matches } = require("../helpers/selectors");
 const DOMException = require("domexception/webidl2js-wrapper");
 const DOMTokenList = require("../generated/DOMTokenList");
 const NamedNodeMap = require("../generated/NamedNodeMap");
-const NodeList = require("../generated/NodeList");
-const idlUtils = require("../generated/utils");
 const validateNames = require("../helpers/validate-names");
 const { asciiLowercase, asciiUppercase } = require("../helpers/strings");
 const { listOfElementsWithQualifiedName, listOfElementsWithNamespaceAndLocalName,
@@ -548,15 +547,6 @@ class ElementImpl extends NodeImpl {
 
   matches(selectors) {
     return matches(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-  }
-
-  querySelector(selectors) {
-    return querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-  }
-
-  querySelectorAll(selectors) {
-    const nodes = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
-    return NodeList.create(this._globalObject, [], { nodes });
   }
 
   webkitMatchesSelector(selectors) {
