+++ lib/jsdom/living/nodes/ParentNode-impl.js
@@ -1,7 +1,10 @@
 "use strict";
 
+const idlUtils = require("../generated/utils");
+const NodeList = require("../generated/NodeList");
 const HTMLCollection = require("../generated/HTMLCollection");
 const { domSymbolTree } = require("../helpers/internal-constants");
+const { querySelector, querySelectorAll } = require("../helpers/selectors");
 const NODE_TYPE = require("../node-type");
 const { convertNodesIntoNode } = require("../node");
 
@@ -56,6 +59,16 @@ class ParentNodeImpl {
     const node = convertNodesIntoNode(this._ownerDocument, nodes);
     this._preInsertValidity(node, null);
     this._replaceAll(node);
+  }
+
+  querySelector(selectors) {
+    return querySelector(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
+  }
+
+  // Warning for internal users: this returns a NodeList containing IDL wrappers instead of impls
+  querySelectorAll(selectors) {
+    const nodes = querySelectorAll(selectors, idlUtils.wrapperForImpl(this), this._globalObject);
+    return NodeList.create(this._globalObject, [], { nodes });
   }
 }
